requirements.txt
Django==3.2.9
djangorestframework
python-dotenv
djangorestframework-simplejwt
drf-writable-nested
requests
django-environ
psycopg2-binary==2.9.2
google-cloud-secret-manager==2.8.0
pygithub
pytest
gitpython
bs4
numpy
scipy
pytest-cov
pytest-json-report
hypothesis
python-dotenv
python-utils